fix: eliminate route user-stats, merge into users
